requirements.txt
crewai
groq
yfinance
pandas
pandas-ta
plotly
streamlit
streamlit-autorefresh
apscheduler
python-telegram-bot
praw
pytrends
feedparser
newsapi-python
requests
beautifulsoup4
pytesseract
Pillow
sqlalchemy
python-dotenv
httpx
